Liaison Back Front

+++ Sufodweb/src/main/java/sufodWeb/restController/PersonnageRestController.java
@@ -4,6 +4,7 @@ import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -11,8 +12,8 @@ import org.springframework.web.bind.annotation.RestController;
 
 import com.fasterxml.jackson.annotation.JsonView;
 
-import sufod.entity.Personnage;
 import sufod.entity.JsonViews;
+import sufod.entity.Personnage;
 import sufod.services.PersonnageService;
 
 @RestController
@@ -30,7 +31,7 @@ public class PersonnageRestController {
 
 	@GetMapping("/{id}")
 	@JsonView(JsonViews.Common.class)
-	public Personnage getById(@PathVariable Long id) {
+	public Personnage getById(@PathVariable Integer id) {
 		return (Personnage) personnageService.getById(id);
 	}
